chore: add check in serve itself
