+++ packages/webpack-cli/lib/commands/ExternalCommand.js
@@ -34,15 +34,6 @@ class ExternalCommand {
                 logger.error(`Action Interrupted, use ${cyan('webpack-cli help')} to see possible commands.`);
             }
         }
-        // Check dev server availability before running serve
-        if (name === 'serve') {
-            try {
-                // eslint-disable-next-line node/no-extraneous-require
-                require.resolve('webpack-dev-server');
-            } catch (error) {
-                throw new Error(`You need to install 'webpack-dev-server' for running 'webpack serve'.\n${error}`);
-            }
-        }
         return pkgLoc ? require(scopeName).default(...args) : null;
     }
 }
